Delete AU-1001 AUG.png
This image causes a Steyr Aug to appear instead of the railgun, I'm not sure if maybe you wanted it that way for a reason, but I downloaded all these files to my PC and replaced the AUG with the railgun I designed and lined it up properly so it would look right, I see that you also have my Railgun in the folder too but when people download the master zip, it comes with the AUG...

If I may ask, could you give my new GitHub account permission to make changes as necessary, I don't intend to spam it or derail or take your project from you, I only want to be able to help and bug fix in my spare time. I won't make any drastic changes, but I would like to be able post my progress to this "GitHub Repository?" for quicker results.